+++ assembl/static2/.storybook/config.js
@@ -8,7 +8,7 @@ import '../css/themes/default/assembl_web.scss';
 addDecorator(centered);
 
 function loadStories() {
-  require('../js/app/integration/101/components/button101.stories.jsx');
+  require('../js/app/integration/101/components/button101/button101.stories.jsx');
 
   // TODO: To clean ?
   // require('../js/app/stories/components/common/common.jsx');
