+++ assembl/static2/css/components/buttons.scss
@@ -222,4 +222,4 @@
 }
 
 // The partial used below is only used as an example for instruction purpose
-@import '../../js/app/integration/101/components/button101';
+@import '../../js/app/integration/101/components/button101/button101';
